requirements.txt
flask
jinja2
flask-wtf
flask-sqlalchemy
mysql-connector
pymysql
requests
flask-migrate